--- bag_unpacker/src/main/scala/uk/ac/wellcome/platform/archive/bagunpacker/config/UnpackerConfig.scala
@@ -1,5 +1,0 @@
-package uk.ac.wellcome.platform.archive.bagunpacker.config
-
-case class UnpackerConfig(
-  bufferSize: Int
-)
