+++ bag_unpacker/src/main/scala/uk/ac/wellcome/platform/archive/bagunpacker/config/builders/UnpackerConfigBuilder.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.archive.bagunpacker.config.builders
 
 import com.typesafe.config.Config
-import uk.ac.wellcome.platform.archive.bagunpacker.config.UnpackerConfig
+import uk.ac.wellcome.platform.archive.bagunpacker.config.models.UnpackerConfig
 import uk.ac.wellcome.typesafe.config.builders.EnrichConfig._
 
 object UnpackerConfigBuilder {
