+++ bag_unpacker/src/main/scala/uk/ac/wellcome/platform/archive/bagunpacker/config/models/UnpackerConfig.scala
@@ -1,0 +1,5 @@
+package uk.ac.wellcome.platform.archive.bagunpacker.config.models
+
+case class UnpackerConfig(
+  bufferSize: Int
+)
